Rework MAC address handling in options
rather than having an option for each mac address
instead we take the --mac or -m argument as the
macid and optionally add the --address or -a
argument that supplies the address to assign to it.

This is the most basic of implementations and should be
expanded on.  providing no mac address should only read
out the address assigned to that id.  Additionally these
options should take a comma separated list of ids and
mac addresses so a bulk update can be done.

Signed-off-by: Jon Nettleton <jon@solid-run.com>